anti juke
So I gotta be honest.
I helped some people with this project, but I had a tiny little secret.

There's no reason to have "n % 10 % 2", instead of "n % 2". I had it in there to be like a safe check. If anybody plagarised my code without understanding what it did, they'd have that in their code. And lo and behold, I found @00100000 with that exact codeblock. So I know he looked through my code. I don't know if I should be proud, or sad.